Resolved all the issues that were there
