+++ Streetcode/Streetcode.BLL/MediatR/Media/Art/Create/CreateArtCommand.cs
@@ -2,7 +2,6 @@
 using MediatR;
 using Streetcode.BLL.DTO.Media.Art;
 
-namespace Streetcode.BLL.MediatR.Media.Art.Create
-{
-    public record CreateArtCommand(ArtCreateRequestDTO ArtCreateRequest) : IRequest<Result<ArtDTO>>;
-}
+namespace Streetcode.BLL.MediatR.Media.Art.Create;
+
+public record CreateArtCommand(ArtCreateRequestDTO ArtCreateRequest) : IRequest<Result<ArtDTO>>;
