+++ Streetcode/Streetcode.BLL/MediatR/Media/Art/Update/UpdateArtCommand.cs
@@ -2,7 +2,6 @@
 using MediatR;
 using Streetcode.BLL.DTO.Media.Art;
 
-namespace Streetcode.BLL.MediatR.Media.Art.Update
-{
-    public record UpdateArtCommand(UpdateArtRequestDTO ArtUpdateRequest) : IRequest<Result<ArtDTO>>;
-}
+namespace Streetcode.BLL.MediatR.Media.Art.Update;
+
+public record UpdateArtCommand(UpdateArtRequestDTO ArtUpdateRequest) : IRequest<Result<ArtDTO>>;
